chore: drop legacy store module
