+++ tests/test_memory_lattice_retrieve.py
@@ -5,11 +5,9 @@ import pytest
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from pro_metrics import tokenize, lowercase  # noqa: E402
-import pro_memory  # noqa: E402
+from memory import storage, lattice  # noqa: E402
 import pro_rag  # noqa: E402
-from memory import MemoryStore  # noqa: E402
 import pro_predict  # noqa: E402
-
 
 def _embed(text: str) -> dict:
     words = lowercase(tokenize(text))
@@ -21,7 +19,7 @@ async def test_retrieve_uses_graph_and_embedding(monkeypatch):
     monkeypatch.setattr(
         pro_predict, "_VECTORS", {"dogs": {"d": 1.0}, "cats": {"c": 1.0}}
     )
-    store = MemoryStore()
+    store = storage.MemoryStore()
     store.add_utterance("d1", "user", "cats eat fish", _embed("cats eat fish"))
     store.add_utterance(
         "d1", "user", "dogs chase cats", _embed("dogs chase cats")
@@ -30,7 +28,7 @@ async def test_retrieve_uses_graph_and_embedding(monkeypatch):
     async def fake_related(words):
         return []
 
-    monkeypatch.setattr(pro_memory, "fetch_related_concepts", fake_related)
+    monkeypatch.setattr(lattice, "fetch_related_concepts", fake_related)
 
     results = await pro_rag.retrieve(["dogs"], lattice=store, limit=2)
     assert results[0] == "dogs chase cats"
