+++ tests/test_memory_vectors.py
@@ -9,34 +9,33 @@ from autograd import numpy as anp
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
-import pro_memory  # noqa: E402
-from memory import MemoryStore  # noqa: E402
+from memory import storage  # noqa: E402
 import morphology  # noqa: E402
 
 
 @pytest.mark.asyncio
 async def test_store_and_fetch_similar_embeddings(tmp_path, monkeypatch):
     db_path = tmp_path / "mem.db"
-    monkeypatch.setattr(pro_memory, "DB_PATH", str(db_path))
-    await pro_memory.init_db()
-    await pro_memory.build_index()
-    await pro_memory.add_message("hello world")
-    await pro_memory.add_message("goodbye world")
-    results = await pro_memory.fetch_similar_messages("hello there", top_k=1)
+    monkeypatch.setattr(storage, "DB_PATH", str(db_path))
+    await storage.init_db()
+    await storage.build_index()
+    await storage.add_message("hello world")
+    await storage.add_message("goodbye world")
+    results = await storage.fetch_similar_messages("hello there", top_k=1)
     assert results == ["hello world"]
     # Update the first embedding and ensure it persists
-    new_vec = np.ones_like(pro_memory._STORE.embeddings[0])
-    await pro_memory.persist_learned_vector(0, new_vec)
+    new_vec = np.ones_like(storage._STORE.embeddings[0])
+    await storage.persist_learned_vector(0, new_vec)
     conn = sqlite3.connect(str(db_path))
     cur = conn.execute("SELECT embedding FROM messages WHERE rowid = 1")
     blob = np.frombuffer(cur.fetchone()[0], dtype=np.float32)
     assert np.allclose(blob, new_vec)
     conn.close()
-    await pro_memory.close_db()
+    await storage.close_db()
 
 
 def test_embedding_storage_and_gradient_retrieval():
-    store = MemoryStore()
+    store = storage.MemoryStore()
     text = "привет мир"
     emb = store.add_utterance("dlg", "user", text)
     dialogue = store.get_dialogue("dlg")
